% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafreq.R
\name{datafreq}
\alias{datafreq}
\title{Frequency Components Decomposition Function}
\usage{
datafreq(data, frequencyPoints = 10000)
}
\arguments{
\item{data}{is an TDMS file Input}

\item{frequencyPoints}{corresponds to the number of points for FFT}
}
\value{
The function returns the plot of the TDMS file data signal in the Frequency Domain
}
\description{
The function reads in a TDMS file and generates its Fast Fourier Transform (FFT). For
the FFT calculation, we ignore the complex conjugates involved and only consider the Magnitudes of each
complex number.
}
