% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demeaner.R
\name{demeaner}
\alias{demeaner}
\title{Data Stream Demeaner}
\usage{
demeaner(datastream)
}
\arguments{
\item{datastream}{is a vector containing a one dimensional data stream}
}
\value{
The function returns a vector containing the demeaned data stream
}
\description{
The function aims to take a data stream in a vector form and returns the
demeaned data for further analysis. In short, the function calculates the mean of the
data and element-wise substracts it from the entire vector
}
\examples{
# Load the installed library/package
library(readTDMS)
# Create or Import the data to be demeaned
data <- c(runif(150))
# Call the Function
demeaned_data <- demeaner(data)
}
