% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmsread.R
\name{tdmsread}
\alias{tdmsread}
\title{Data Stream (.tdms Format) Reading Function}
\usage{
tdmsread(file_input, plot = FALSE)
}
\arguments{
\item{file_input}{is an TDMS file Input}

\item{plot}{is a boolean indicating if plotting is desired}
}
\value{
The function returns a list containing:
\itemize{
\item datastream - A vector of data stream values based on the input TDMS file
\item timestamps - A vector of time stamps associated with the datastream based on the input TDMS file
\item summ - The summary of the extracted data stream
\item img - The Time domain plot of the data stream
}
}
\description{
The function reads in an TDMS file as an input.
The destination/folder of the file must be specified and its associated
TDMS_Index file must be present in the same folder.
}
\examples{
# Load the installed library/package
library(readTDMS)
# Open the TDMS File with specified path
f <- file('~/readTDMS/vignettes/data/file.tdms', 'rb')
# Call the Function to get the required plot
ans <- tdmsread(f)
# Datastream values
ans$datastream
# Timestamps
ans$timestamps
# Close the file if needed
# Note: For any changes and rerun of about function, we need to reopen the TDMS file
close(f)
}
