% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavfft.R
\name{wavfft}
\alias{wavfft}
\title{Data Stream FFT Function}
\usage{
wavfft(data, plot = TRUE)
}
\arguments{
\item{data}{is the data stream on which the FFT is applied}

\item{plot}{indicates whether the user wants to plot the data in the frequency domain or NOT. By default, it is TRUE
results in a plot representing the frequency domain of the data stream with various
frequency components}
}
\value{
The function returns a list containing :
\itemize{
\item data_FFT - A vector of frequency components associated with the data stream
\item frequency - A vector of frequencies associated with FFT on data stream
\item img - A plot of the data in the frequency domain
}
}
\description{
This function decomposes the data stream into its frequency components using the
Fast Fourier Transform (FFT). Simply calling default FFT function in R does not give intuitive
results so this function also transforms the output we get
}
