% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavread.R
\name{wavread}
\alias{wavread}
\title{Data Stream (.wav Format) Reading Function}
\usage{
wavread(file_input)
}
\arguments{
\item{file_input}{is an Wav file Input}
}
\value{
The function returns a list containing:
\itemize{
\item data - A matrix or vector containing the data from the input .wav or Audio file
\item signal - A vector representing the extracted signal (demeaned)
\item samplingRate - A scalar corresponding to the Sampling Rate of the Signal
\item duration - A scalar corresponding to the Duration of the Signal
}
}
\description{
The function reads in an Wav file as an input and extracts the data stream and signal
as a matrix and vector respectively along with the total duration and sampling rate of the stream
}
