% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavsumm.R
\name{wavsumm}
\alias{wavsumm}
\title{Data Stream (.wav Format) Summary and Plotting Function}
\usage{
wavsumm(file_input, plot = TRUE)
}
\arguments{
\item{file_input}{is an Wav file Input}

\item{plot}{A boolean which indicates if we want to plot the signal or not. It is TRUE by default.}
}
\value{
The function returns a list containing:
\itemize{
\item summ - The summary of the extracted data stream
\item img - The plot of the data stream if plotting is specified as TRUE
}
}
\description{
The function reads in an Wav file as an input and extracts the data stream from it. It
also calculations the summary of the data stream and plots it if specified.
}
