% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafreq.R
\name{datafreq}
\alias{datafreq}
\title{Frequency Components Decomposition Function}
\usage{
datafreq(
  file_input,
  frequencyPoints = 10000,
  xlim1 = 0,
  xlim2 = 10000,
  ylim1 = -10,
  ylim2 = 500
)
}
\arguments{
\item{file_input}{is an TDMS file Input}

\item{frequencyPoints}{corresponds to the number of points for FFT (by default = 10000) (scalar)}

\item{xlim1}{specifies the lower limit for x-axis of the plot (by default = 0) (scalar)}

\item{xlim2}{specifies the upper limit for x-axis of the plot (by default = 10000) (scalar)}

\item{ylim1}{specifies the lower limit for y-axis of the plot (by default = -10) (scalar)}

\item{ylim2}{specifies the upper limit for y-axis of the plot (by default = 500) (scalar)}
}
\value{
The function returns a list containing :
\itemize{
\item data_FFT - A vector of frequency components associated with the data stream
\item frequency - A vector of frequencies associated with FFT on data stream
\item img - A plot of the TDMS file data signal in the frequency domain
}
}
\description{
The function reads in a TDMS file and generates its Fast Fourier Transform (FFT). For
the FFT calculation, we ignore the complex conjugates involved and only consider the Magnitudes of each
complex number.
}
\examples{
# Load the installed library/package
library(readTDMS)
# Open the TDMS File with specified path
f <- file('~/readTDMS/vignettes/data/file.tdms', 'rb')
# Specify required parameters
frequencyPoints <- 5000
xlim1 <- -200
xlim2 <- 1200
ylim1 <- -10
ylim2 <- 400
# Call the Function to get the required plot
ans <- datafreq(f, frequencyPoints = frequencyPoints, xlim1 = xlim1, xlim2 = xlim2, ylim1 = ylim1, ylim2 = ylim2)
# Close the file if needed
# Note: For any changes and rerun of about function, we need to reopen the TDMS file
close(f)
}
