% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftprofile.R
\name{fftprofile}
\alias{fftprofile}
\title{Top Frequency Component Finder Function}
\usage{
fftprofile(data, frequencyPoints = 10000, n)
}
\arguments{
\item{data}{is an TDMS file Input}

\item{frequencyPoints}{corresponds to the number of points for FFT}

\item{n}{corresponds to the number of Top Frequencies to be returned}
}
\value{
a list of Top frequencies associated with the TDMS file data signal in the Frequency Domain
}
\description{
The function reads in a TDMS file and generates its Fast Fourier Transform (FFT). For
the FFT calculation, we ignore the complex conjugates involved and only consider the Magnitudes of each
complex number. After finding the frequency components, it returns a list of top frequencies based on
specified n value
}
\examples{
# Load the installed library/package
library(readTDMS)

}
