% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmsread.R
\name{tdmsread}
\alias{tdmsread}
\title{Data Stream (.tdms Format) Reading Function}
\usage{
tdmsread(file_input)
}
\arguments{
\item{file_input}{is an TDMS file Input}
}
\value{
The function returns a list containing:
\itemize{
\item datastream - A vector of data stream values based on the input TDMS file
\item timestamps - A vector of time stamps associated with the datastream based on the input TDMS file
}
}
\description{
The function reads in an TDMS file as an input.
The destination/folder of the file must be specified and its associated
TDMS_Index file must be present in the same folder.
}
